cloud/services/analytic/services/analytics-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
confluent-kafka==2.4.0
SQLAlchemy==2.0.35
psycopg[binary,pool]==3.2.1
prometheus-client==0.20.0
confluent-kafka==2.4.0
apscheduler==3.10.4
python-dotenv==1.0.1
# apscheduler==3.10.4
# opentelemetry-sdk==1.26.0
# opentelemetry-exporter-otlp==1.26.0
# opentelemetry-instrumentation-fastapi==0.47b0